关于issue#62的修改

+++ app/models/kc_courses/concerns/ware_reading_method.rb
@@ -10,7 +10,7 @@ module KcCourses
 
       # 设置 user 已经完成了 ware 百分之 read_percent 的内容
       def set_read_percent_by_user(user, read_percent)
-        return true if read_percent_of_user(user) >= (read_percent).to_i
+        return true if read_percent_of_user(user) >= (read_percent).to_f
         ware_reading = self.ware_readings.where(:creator_id => user.id.to_s).first
         if ware_reading.blank?
           self.ware_readings.create(:creator => user, :read_percent => read_percent)
